refactor demo code fo TS, adding mocha types
