+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "nodemon source/server.ts",
     "build": "rm -rf build/ && prettier --write source/ && tsc",
-    "testing": "mocha ./source/qa_testing/check.test.js"
+    "testing": "mocha ./source/test/demo.test.ts"
   },
   "repository": {
     "type": "git",
@@ -34,5 +34,8 @@
     "nodemon": "2.0.20",
     "ts-node": "10.9.1",
     "typescript": "4.8.3"
+  },
+  "devDependencies": {
+    "@types/mocha": "^10.0.0"
   }
 }
